services/service3/requirements.txt
flask
requests
pillow
numpy
python-dotenv